test(security): make #297 command-injection tests argv-aware
The #297 regression tests still reported `xfailed` long after the fix landed,
so they silently claimed the injection was live. `build_hashcat_command` now
returns an argv LIST, but the assertions were `assert payload not in cmd` —
against a list that is a membership check, and a correctly-passed payload IS
an element, so the assertion failed precisely because the code was safe.

Replace test_command_injection_xfail.py with test_command_injection_argv.py:
same payload matrix (6 shell payloads x modes 3/6/7 for the mask, plus the
quote-break-out payloads for -j/-k), now asserting argv safety — the payload is
exactly one whole element, is never spliced into a larger token, and no other
token carries shell metacharacters. No xfail markers, so reintroducing string
concatenation fails the suite (verified by mutation: joining the mask back into
the target token fails all six mode-3 cases). Also adds coverage for a chunk
sub-mask, which the old file never exercised.

Update TESTING.md and the test_security_hardening.py cross-reference: #297 is
fixed, #298 (CSRF) remains the only open strict-xfail finding. Both docs record
the `payload not in argv` trap so it is not reintroduced.
